Split instrumentation from variables
